refactor: use shared gerichtskostenFromBetrag
